Rearrange modules, apply stylish-haskell and fix dropDocument
