+++ src/ArangoDB.hs
@@ -1,1 +1,8 @@
-module ArangoDB where
+module ArangoDB (
+  module ArangoDB.Collections,
+  module ArangoDB.Documents,
+) where
+
+import           ArangoDB.Collections
+import           ArangoDB.Documents
+
