+++ test/doctests.hs
@@ -1,8 +1,8 @@
 module Main where
 
-import Build_doctests (flags, pkgs, module_sources)
-import Data.Foldable (traverse_)
-import Test.DocTest
+import           Build_doctests (flags, module_sources, pkgs)
+import           Data.Foldable  (traverse_)
+import           Test.DocTest
 
 main :: IO ()
 main = do
